requirements.txt
Flask==2.3.3
pydub==0.25.1
openai-whisper==20231117
ffmpeg-python==0.2.0
torch==2.2.2
numpy==1.26.4